Add a workaround for GPU nodes failing during inspection
The NIC sends a different DHCP client ID than the inspection ramdisk which
causes dnsmasq to fail to provide a lease to the ramdisk.  We add a DHCP host
rule in the ironic inspector dnsmasq configuration that causes dnsmasq to use
the MAC address rather than client ID of these hosts to index the lease
database.