Enable Kafka on the monitoring server
This uses the version of Kafka currently provided by LXC, as
more recent versions require Monasca services to be upgraded. When
the Monasca services are upgraded we can go back to the listeners
field and automatic broker id generation.

The host.name field is set to the API IP address, because
otherwise it defaults to the hostname, and since DNS is not
configured the Monasca-API can't talk to it.

This implicitly enables Zookeeper.