Fix kayobe network connectivity check
The following error was observed whilst runing kayobe network connectivity check:

failed: [sv-b16-u21] (item=alaska_ctl) => {
    "changed": false,
    "cmd": "ping  -c1 -M do  # localhost",
    "delta": "0:00:00.003493",
    "end": "2020-05-29 13:39:16.911225",
    "invocation": {
        "module_args": {
            "_raw_params": "ping  -c1 -M do  # localhost",

NOTE: I have addded the `#localhost` comment for debugging purposes it corresponds
to the variable, `remote_host`.

The problem was that the IPs for the seed hypervisor were not in network-allocations.
I've also renamed it to stop it being confused with the implicit local host.

This did not happen on every run as we use the random filter to pick a host.